move welcome screen to app directory
